refactor: move MoneyContent to commonMain
Splits MoneyContent out of MoneyScreen.kt into commonMain so it can be
exercised by PreviewScreenshotGeneratorTest on the JVM target.
