+++ feature/money/src/wasmJsMain/kotlin/feature/money/MoneyScreen.kt
@@ -1,43 +1,7 @@
 package feature.money
 
-import androidx.compose.foundation.gestures.detectTapGestures
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.automirrored.filled.ArrowForward
-import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.Delete
-import androidx.compose.material.icons.filled.Edit
-import androidx.compose.material.icons.filled.Warning
-import androidx.compose.material3.*
-import androidx.compose.runtime.*
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.input.pointer.pointerInput
-import androidx.compose.ui.platform.LocalFocusManager
-import androidx.compose.ui.text.SpanStyle
-import androidx.compose.ui.text.buildAnnotatedString
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.withStyle
-import androidx.compose.ui.unit.dp
+import androidx.compose.runtime.Composable
 import core.ui.LocalWindowSizeClass
-import core.ui.WindowSizeClass
-import core.ui.extensions.displayName
-import core.ui.extensions.icon
-import core.ui.formatYen
-import model.MoneyItem
-import model.MoneyTags
-import model.MonthlyMoney
-import model.MonthlyMoneyStatus
-import model.Payment
-import model.Share
-import model.User
 import org.koin.compose.viewmodel.koinViewModel
 
 @Composable
@@ -64,885 +28,5 @@ fun MoneyScreen(vm: MoneyViewModel = koinViewModel()) {
         onUpdateStatus = vm::onUpdateStatus,
         onImportRecurringItems = vm::onImportRecurringItems,
         windowSizeClass = windowSizeClass,
-    )
-}
-
-@Composable
-internal fun MoneyContent(
-    monthlyMoney: MonthlyMoney,
-    currentYearMonth: String,
-    loading: Boolean,
-    saving: Boolean,
-    statusSaving: Boolean,
-    error: String?,
-    users: List<User>,
-    editingItem: MoneyItem?,
-    formKey: Int,
-    onPreviousMonth: () -> Unit,
-    onNextMonth: () -> Unit,
-    onEditItem: (MoneyItem) -> Unit,
-    onClearForm: () -> Unit,
-    onDeleteItem: (MoneyItem) -> Unit,
-    onMoveItem: (MoneyItem, Int) -> Unit,
-    onSaveItem: (String, Long, String, List<Share>, List<String>) -> Unit,
-    onUpdateStatus: (MonthlyMoneyStatus) -> Unit,
-    onImportRecurringItems: () -> Unit,
-    windowSizeClass: WindowSizeClass = WindowSizeClass.Expanded,
-) {
-    val status = monthlyMoney.status
-    val frozen = status == MonthlyMoneyStatus.FROZEN
-    val isCompact = windowSizeClass == WindowSizeClass.Compact
-    // Compact 用: フォーム表示切替
-    var showFormCompact by remember { mutableStateOf(false) }
-    val focusManager = LocalFocusManager.current
-
-    // editingItem がクリアされたらフォームを閉じる（保存成功時）
-    LaunchedEffect(editingItem) {
-        if (editingItem == null) showFormCompact = false
-    }
-
-    Box(
-        modifier =
-            Modifier
-                .fillMaxSize()
-                .pointerInput(Unit) {
-                    detectTapGestures(onTap = { focusManager.clearFocus() })
-                },
-    ) {
-        if (isCompact) {
-            if (showFormCompact) {
-                // Compact: フォーム表示
-                Column(
-                    modifier =
-                        Modifier
-                            .fillMaxSize()
-                            .padding(12.dp),
-                ) {
-                    Text(
-                        text = if (editingItem != null) "項目を編集" else "項目を追加",
-                        style = MaterialTheme.typography.headlineSmall,
-                        color = MaterialTheme.colorScheme.primary,
-                    )
-                    Spacer(modifier = Modifier.height(8.dp))
-                    MoneyItemForm(
-                        item = editingItem,
-                        formKey = formKey,
-                        users = users,
-                        saving = saving,
-                        frozen = frozen,
-                        onSave = onSaveItem,
-                        onCancel = {
-                            onClearForm()
-                            showFormCompact = false
-                        },
-                        modifier = Modifier.fillMaxWidth().weight(1f),
-                    )
-                }
-            } else {
-                // Compact: リスト表示
-                Column(
-                    modifier =
-                        Modifier
-                            .fillMaxSize()
-                            .padding(12.dp),
-                ) {
-                    Text(
-                        text = "お金の管理",
-                        style = MaterialTheme.typography.headlineSmall,
-                        color = MaterialTheme.colorScheme.primary,
-                    )
-                    Spacer(modifier = Modifier.height(8.dp))
-                    MonthSelector(
-                        yearMonth = currentYearMonth,
-                        onPrevious = onPreviousMonth,
-                        onNext = onNextMonth,
-                    )
-                    Spacer(modifier = Modifier.height(8.dp))
-                    MonthStatusSelector(
-                        status = status,
-                        onStatusChange = onUpdateStatus,
-                        enabled = !statusSaving,
-                    )
-                    Spacer(modifier = Modifier.height(8.dp))
-
-                    // 追加ボタン + インポートボタン
-                    if (!loading && error == null) {
-                        Button(
-                            onClick = {
-                                onClearForm()
-                                showFormCompact = true
-                            },
-                            modifier = Modifier.fillMaxWidth(),
-                            enabled = !frozen,
-                        ) {
-                            Icon(Icons.Default.Add, contentDescription = null)
-                            Spacer(modifier = Modifier.width(8.dp))
-                            Text("項目を追加")
-                        }
-                        if (!frozen) {
-                            Spacer(modifier = Modifier.height(4.dp))
-                            OutlinedButton(
-                                onClick = onImportRecurringItems,
-                                modifier = Modifier.fillMaxWidth(),
-                                enabled = !saving,
-                            ) {
-                                ImportButtonText()
-                            }
-                        }
-                        Spacer(modifier = Modifier.height(8.dp))
-                    }
-
-                    MoneyListContent(
-                        monthlyMoney = monthlyMoney,
-                        loading = loading,
-                        error = error,
-                        users = users,
-                        isCompact = true,
-                        frozen = frozen,
-                        onEditItem = { item ->
-                            onEditItem(item)
-                            showFormCompact = true
-                        },
-                        onDeleteItem = onDeleteItem,
-                        onMoveItem = onMoveItem,
-                        modifier = Modifier.weight(1f),
-                    )
-                }
-            }
-        } else {
-            // Medium/Expanded: 左にリスト、右にフォーム（常時表示）
-            Column(
-                modifier =
-                    Modifier
-                        .fillMaxSize()
-                        .padding(24.dp),
-            ) {
-                Text(
-                    text = "お金の管理",
-                    style = MaterialTheme.typography.headlineLarge,
-                    color = MaterialTheme.colorScheme.primary,
-                )
-                Spacer(modifier = Modifier.height(16.dp))
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(12.dp),
-                ) {
-                    MonthSelector(
-                        yearMonth = currentYearMonth,
-                        onPrevious = onPreviousMonth,
-                        onNext = onNextMonth,
-                    )
-                    MonthStatusSelector(
-                        status = status,
-                        onStatusChange = onUpdateStatus,
-                        enabled = !statusSaving,
-                    )
-                    if (!frozen && !loading && error == null) {
-                        OutlinedButton(
-                            onClick = onImportRecurringItems,
-                            enabled = !saving,
-                        ) {
-                            ImportButtonText()
-                        }
-                    }
-                }
-                Spacer(modifier = Modifier.height(16.dp))
-                Row(
-                    modifier = Modifier.fillMaxWidth().weight(1f),
-                ) {
-                    MoneyListContent(
-                        monthlyMoney = monthlyMoney,
-                        loading = loading,
-                        error = error,
-                        users = users,
-                        isCompact = false,
-                        frozen = frozen,
-                        onEditItem = onEditItem,
-                        onDeleteItem = onDeleteItem,
-                        onMoveItem = onMoveItem,
-                        modifier = Modifier.weight(1f),
-                    )
-
-                    Spacer(modifier = Modifier.width(24.dp))
-
-                    MoneyItemForm(
-                        item = editingItem,
-                        formKey = formKey,
-                        users = users,
-                        saving = saving,
-                        frozen = frozen,
-                        onSave = onSaveItem,
-                        onCancel = onClearForm,
-                        modifier = Modifier.width(400.dp),
-                    )
-                }
-            }
-        }
-
-        // 保存中インジケーター
-        if (saving) {
-            LinearProgressIndicator(modifier = Modifier.fillMaxWidth().align(Alignment.TopCenter))
-        }
-    }
-}
-
-@Composable
-private fun MoneyListContent(
-    monthlyMoney: MonthlyMoney,
-    loading: Boolean,
-    error: String?,
-    users: List<User>,
-    isCompact: Boolean,
-    frozen: Boolean,
-    onEditItem: (MoneyItem) -> Unit,
-    onDeleteItem: (MoneyItem) -> Unit,
-    onMoveItem: (MoneyItem, Int) -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    when {
-        loading -> {
-            Box(
-                modifier = modifier.fillMaxWidth(),
-                contentAlignment = Alignment.Center,
-            ) {
-                CircularProgressIndicator()
-            }
-        }
-
-        error != null -> {
-            Box(modifier = modifier) {
-                Text("エラー: $error", color = MaterialTheme.colorScheme.error)
-            }
-        }
-
-        else -> {
-            val spacing = if (isCompact) 8.dp else 12.dp
-            LazyColumn(
-                modifier = modifier.fillMaxWidth(),
-                verticalArrangement = Arrangement.spacedBy(spacing),
-                contentPadding = PaddingValues(bottom = 16.dp),
-            ) {
-                item(key = "summary") {
-                    SummaryCard(
-                        items = monthlyMoney.items,
-                        users = users,
-                        payments = monthlyMoney.payments,
-                        isCompact = isCompact,
-                    )
-                }
-
-                if (monthlyMoney.items.isEmpty()) {
-                    item(key = "empty") {
-                        Box(
-                            modifier = Modifier.fillMaxWidth().padding(32.dp),
-                            contentAlignment = Alignment.Center,
-                        ) {
-                            Text(
-                                text = "この月のデータはありません",
-                                style = MaterialTheme.typography.bodyLarge,
-                                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                            )
-                        }
-                    }
-                }
-
-                val sortedItems = monthlyMoney.items.sortedByDescending { it.tags.isNotEmpty() }
-                items(sortedItems, key = { it.id }) { item ->
-                    MoneyItemCard(
-                        item = item,
-                        users = users,
-                        onEdit = { onEditItem(item) },
-                        onDelete = { onDeleteItem(item) },
-                        onMovePrev = { onMoveItem(item, -1) },
-                        onMoveNext = { onMoveItem(item, 1) },
-                        isCompact = isCompact,
-                        frozen = frozen,
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun MoneyItemForm(
-    item: MoneyItem?,
-    formKey: Int,
-    users: List<User>,
-    saving: Boolean,
-    frozen: Boolean,
-    onSave: (String, Long, String, List<Share>, List<String>) -> Unit,
-    onCancel: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    // formKey をキーに含めて、clearForm 時に確実にリセットする
-    val key = item ?: formKey
-    var name by remember(key) { mutableStateOf(item?.name ?: "") }
-    var amountText by remember(key) { mutableStateOf(item?.amount?.toString() ?: "") }
-    var note by remember(key) { mutableStateOf(item?.note ?: "") }
-    var selectedTags by remember(key) { mutableStateOf(item?.tags ?: emptyList()) }
-    var shareAmounts by remember(key) {
-        mutableStateOf(
-            users.associate { user ->
-                user.uid to (
-                    item
-                        ?.shares
-                        ?.find { it.uid == user.uid }
-                        ?.amount
-                        ?.toString() ?: ""
-                )
-            },
-        )
-    }
-
-    val parsedAmount = amountText.toLongOrNull()
-    val isAmountValid = parsedAmount != null
-    val amount = parsedAmount ?: 0L
-    val shares =
-        shareAmounts.mapNotNull { (uid, text) ->
-            val a = text.toLongOrNull()
-            if (a != null && a != 0L) Share(uid, a) else null
-        }
-    val shareTotal = shares.sumOf { it.amount }
-    val mismatch = shares.isNotEmpty() && shareTotal != amount
-    val isEditing = item != null
-
-    Card(
-        modifier = modifier,
-        colors =
-            CardDefaults.cardColors(
-                containerColor = MaterialTheme.colorScheme.surfaceVariant,
-            ),
-    ) {
-        Column(
-            modifier =
-                Modifier
-                    .fillMaxWidth()
-                    .verticalScroll(rememberScrollState())
-                    .padding(20.dp),
-            verticalArrangement = Arrangement.spacedBy(12.dp),
-        ) {
-            Text(
-                text = if (isEditing) "項目を編集" else "項目を追加",
-                style = MaterialTheme.typography.titleLarge,
-            )
-
-            OutlinedTextField(
-                value = name,
-                onValueChange = { name = it },
-                label = { Text("項目名") },
-                modifier = Modifier.fillMaxWidth(),
-                singleLine = true,
-                enabled = !saving,
-            )
-
-            OutlinedTextField(
-                value = amountText,
-                onValueChange = { amountText = it.filter { c -> c.isDigit() || c == '-' } },
-                label = { Text("金額 (円)") },
-                modifier = Modifier.fillMaxWidth(),
-                singleLine = true,
-                keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
-                enabled = !saving,
-            )
-
-            OutlinedTextField(
-                value = note,
-                onValueChange = { note = it },
-                label = { Text("備考") },
-                modifier = Modifier.fillMaxWidth(),
-                minLines = 1,
-                maxLines = 3,
-                enabled = !saving,
-            )
-
-            FlowRow(
-                horizontalArrangement = Arrangement.spacedBy(8.dp),
-            ) {
-                val allTags = listOf(MoneyTags.RECURRING, MoneyTags.CARRY_OVER)
-                allTags.forEach { tag ->
-                    FilterChip(
-                        selected = tag in selectedTags,
-                        onClick = {
-                            selectedTags =
-                                if (tag in selectedTags) selectedTags - tag else selectedTags + tag
-                        },
-                        label = { Text(tag) },
-                        enabled = !saving,
-                    )
-                }
-            }
-
-            if (users.isNotEmpty()) {
-                Text(
-                    text = "支払い分担",
-                    style = MaterialTheme.typography.titleSmall,
-                )
-
-                for (user in users) {
-                    Row(
-                        verticalAlignment = Alignment.CenterVertically,
-                        horizontalArrangement = Arrangement.spacedBy(8.dp),
-                    ) {
-                        OutlinedTextField(
-                            value = shareAmounts[user.uid] ?: "",
-                            onValueChange = { value ->
-                                shareAmounts =
-                                    shareAmounts.toMutableMap().apply {
-                                        put(user.uid, value.filter { c -> c.isDigit() || c == '-' })
-                                    }
-                            },
-                            label = { Text(user.displayName ?: user.uid) },
-                            modifier = Modifier.weight(1f),
-                            singleLine = true,
-                            keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
-                            suffix = { Text("円") },
-                            enabled = !saving,
-                        )
-                        OutlinedButton(
-                            onClick = {
-                                shareAmounts =
-                                    shareAmounts.toMutableMap().apply {
-                                        put(user.uid, (amount / 2).toString())
-                                    }
-                            },
-                            enabled = !saving && isAmountValid,
-                            contentPadding = PaddingValues(horizontal = 8.dp),
-                        ) {
-                            Text("半額", style = MaterialTheme.typography.labelSmall)
-                        }
-                        OutlinedButton(
-                            onClick = {
-                                shareAmounts =
-                                    shareAmounts.toMutableMap().apply {
-                                        put(user.uid, amount.toString())
-                                    }
-                            },
-                            enabled = !saving && isAmountValid,
-                            contentPadding = PaddingValues(horizontal = 8.dp),
-                        ) {
-                            Text("全額", style = MaterialTheme.typography.labelSmall)
-                        }
-                        OutlinedButton(
-                            onClick = {
-                                val othersTotal =
-                                    shareAmounts
-                                        .filter { it.key != user.uid }
-                                        .values
-                                        .sumOf { it.toLongOrNull() ?: 0L }
-                                shareAmounts =
-                                    shareAmounts.toMutableMap().apply {
-                                        put(user.uid, (amount - othersTotal).toString())
-                                    }
-                            },
-                            enabled = !saving && isAmountValid,
-                            contentPadding = PaddingValues(horizontal = 8.dp),
-                        ) {
-                            Text("残額", style = MaterialTheme.typography.labelSmall)
-                        }
-                    }
-                }
-
-                if (mismatch) {
-                    Row(
-                        verticalAlignment = Alignment.CenterVertically,
-                        horizontalArrangement = Arrangement.spacedBy(4.dp),
-                    ) {
-                        Icon(
-                            Icons.Default.Warning,
-                            contentDescription = null,
-                            tint = MaterialTheme.colorScheme.error,
-                            modifier = Modifier.size(16.dp),
-                        )
-                        Text(
-                            text = "差額: ${formatYen(shareTotal - amount)}",
-                            style = MaterialTheme.typography.bodySmall,
-                            color = MaterialTheme.colorScheme.error,
-                        )
-                    }
-                }
-            }
-
-            if (frozen) {
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(4.dp),
-                ) {
-                    Icon(
-                        MonthlyMoneyStatus.FROZEN.icon,
-                        contentDescription = null,
-                        tint = MaterialTheme.colorScheme.error,
-                        modifier = Modifier.size(16.dp),
-                    )
-                    Text(
-                        text = "この月は凍結されています",
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.error,
-                    )
-                }
-            }
-
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.End,
-                verticalAlignment = Alignment.CenterVertically,
-            ) {
-                TextButton(onClick = onCancel, enabled = !saving) {
-                    Text("キャンセル")
-                }
-                Spacer(modifier = Modifier.width(8.dp))
-                Button(
-                    onClick = { onSave(name, amount, note, shares, selectedTags) },
-                    enabled = name.isNotBlank() && isAmountValid && !saving && !frozen,
-                ) {
-                    Text(if (isEditing) "保存" else "追加")
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun MonthSelector(
-    yearMonth: String,
-    onPrevious: () -> Unit,
-    onNext: () -> Unit,
-) {
-    val parts = yearMonth.split("-")
-    val displayText = "${parts[0]}年${parts[1].toInt()}月"
-
-    Row(
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(8.dp),
-    ) {
-        IconButton(onClick = onPrevious) {
-            Icon(Icons.AutoMirrored.Filled.ArrowBack, contentDescription = "前月")
-        }
-        Text(
-            text = displayText,
-            style = MaterialTheme.typography.titleLarge,
-        )
-        IconButton(onClick = onNext) {
-            Icon(Icons.AutoMirrored.Filled.ArrowForward, contentDescription = "翌月")
-        }
-    }
-}
-
-@Composable
-private fun MonthStatusSelector(
-    status: MonthlyMoneyStatus,
-    onStatusChange: (MonthlyMoneyStatus) -> Unit,
-    enabled: Boolean = true,
-    modifier: Modifier = Modifier,
-) {
-    FlowRow(
-        modifier = modifier,
-        horizontalArrangement = Arrangement.spacedBy(6.dp),
-        verticalArrangement = Arrangement.spacedBy(4.dp),
-    ) {
-        for (candidate in MonthlyMoneyStatus.entries) {
-            val selected = candidate == status
-            FilterChip(
-                selected = selected,
-                onClick = { if (!selected) onStatusChange(candidate) },
-                label = { Text(candidate.displayName, style = MaterialTheme.typography.labelMedium) },
-                enabled = enabled,
-                leadingIcon = {
-                    Icon(
-                        candidate.icon,
-                        contentDescription = null,
-                        modifier = Modifier.size(16.dp),
-                    )
-                },
-            )
-        }
-    }
-}
-
-@Composable
-private fun SummaryCard(
-    items: List<MoneyItem>,
-    users: List<User>,
-    payments: List<Payment>,
-    isCompact: Boolean,
-) {
-    val totalAmount = items.sumOf { it.amount }
-
-    // ユーザーごとの割当合計
-    val userAllocated = mutableMapOf<String, Long>()
-    for (item in items) {
-        for (share in item.shares) {
-            userAllocated[share.uid] = (userAllocated[share.uid] ?: 0L) + share.amount
-        }
-    }
-
-    // ユーザーごとの支払い済み合計
-    val userPaid = mutableMapOf<String, Long>()
-    for (payment in payments) {
-        userPaid[payment.uid] = (userPaid[payment.uid] ?: 0L) + payment.amount
-    }
-
-    // 全ユーザーが割当を満たしているかチェック
-    val allPaid =
-        userAllocated.isNotEmpty() &&
-            userAllocated.all { (uid, allocated) ->
-                (userPaid[uid] ?: 0L) >= allocated
-            }
-
-    Card(
-        modifier =
-            Modifier.fillMaxWidth().let {
-                if (!isCompact) it.widthIn(max = 600.dp) else it
-            },
-        colors =
-            CardDefaults.cardColors(
-                containerColor = MaterialTheme.colorScheme.surfaceVariant,
-            ),
-    ) {
-        Column(modifier = Modifier.padding(20.dp)) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically,
-            ) {
-                Column {
-                    Text(
-                        text = "合計",
-                        style = MaterialTheme.typography.titleMedium,
-                    )
-                    Text(
-                        text = formatYen(totalAmount),
-                        style = MaterialTheme.typography.headlineMedium,
-                        color = MaterialTheme.colorScheme.primary,
-                    )
-                }
-                if (allPaid) {
-                    Surface(
-                        color = MaterialTheme.colorScheme.primary,
-                        shape = MaterialTheme.shapes.small,
-                    ) {
-                        Text(
-                            text = "支払い完了",
-                            modifier = Modifier.padding(horizontal = 12.dp, vertical = 6.dp),
-                            style = MaterialTheme.typography.labelLarge,
-                            color = MaterialTheme.colorScheme.onPrimary,
-                        )
-                    }
-                }
-            }
-
-            val totalAllocated = userAllocated.values.sum()
-            val allocationMismatch = userAllocated.isNotEmpty() && totalAllocated != totalAmount
-
-            if (allocationMismatch) {
-                Spacer(modifier = Modifier.height(4.dp))
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(4.dp),
-                ) {
-                    Icon(
-                        Icons.Default.Warning,
-                        contentDescription = null,
-                        tint = MaterialTheme.colorScheme.error,
-                        modifier = Modifier.size(16.dp),
-                    )
-                    Text(
-                        text = "差額: ${formatYen(totalAllocated - totalAmount)}",
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.error,
-                    )
-                }
-            }
-
-            if (userAllocated.isNotEmpty()) {
-                Spacer(modifier = Modifier.height(8.dp))
-                HorizontalDivider(color = MaterialTheme.colorScheme.outlineVariant)
-                Spacer(modifier = Modifier.height(8.dp))
-
-                for ((uid, allocated) in userAllocated) {
-                    val userName = users.find { it.uid == uid }?.displayName ?: uid.take(8)
-                    val paid = userPaid[uid] ?: 0L
-                    Row(
-                        modifier = Modifier.fillMaxWidth(),
-                        horizontalArrangement = Arrangement.SpaceBetween,
-                    ) {
-                        Text(
-                            text = userName,
-                            style = MaterialTheme.typography.bodyMedium,
-                        )
-                        Text(
-                            text = "${formatYen(paid)} / ${formatYen(allocated)}",
-                            style = MaterialTheme.typography.bodyMedium,
-                            color =
-                                if (paid >= allocated) {
-                                    MaterialTheme.colorScheme.primary
-                                } else {
-                                    MaterialTheme.colorScheme.onSurfaceVariant
-                                },
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun MoneyItemCard(
-    item: MoneyItem,
-    users: List<User>,
-    onEdit: () -> Unit,
-    onDelete: () -> Unit,
-    onMovePrev: () -> Unit,
-    onMoveNext: () -> Unit,
-    isCompact: Boolean,
-    frozen: Boolean,
-) {
-    val shareTotal = item.shares.sumOf { it.amount }
-    val mismatch = shareTotal != item.amount && item.shares.isNotEmpty()
-
-    Card(
-        modifier =
-            Modifier.fillMaxWidth().let {
-                if (!isCompact) it.widthIn(max = 600.dp) else it
-            },
-        colors =
-            CardDefaults.cardColors(
-                containerColor = MaterialTheme.colorScheme.surfaceVariant,
-            ),
-    ) {
-        Column(modifier = Modifier.padding(20.dp)) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically,
-            ) {
-                Column(modifier = Modifier.weight(1f)) {
-                    Row(
-                        verticalAlignment = Alignment.CenterVertically,
-                        horizontalArrangement = Arrangement.spacedBy(8.dp),
-                    ) {
-                        Text(
-                            text = item.name,
-                            style = MaterialTheme.typography.titleMedium,
-                        )
-                        for (tag in item.tags) {
-                            Surface(
-                                color = MaterialTheme.colorScheme.secondaryContainer,
-                                shape = MaterialTheme.shapes.small,
-                            ) {
-                                Text(
-                                    text = tag,
-                                    modifier = Modifier.padding(horizontal = 6.dp, vertical = 2.dp),
-                                    style = MaterialTheme.typography.labelSmall,
-                                    color = MaterialTheme.colorScheme.onSecondaryContainer,
-                                )
-                            }
-                        }
-                    }
-                    Text(
-                        text = formatYen(item.amount),
-                        style = MaterialTheme.typography.headlineSmall,
-                        color = MaterialTheme.colorScheme.primary,
-                    )
-                }
-
-                Row {
-                    if (!frozen) {
-                        IconButton(onClick = onMovePrev) {
-                            Icon(
-                                Icons.AutoMirrored.Filled.ArrowBack,
-                                contentDescription = "前月へ移動",
-                                tint = MaterialTheme.colorScheme.onSurfaceVariant,
-                            )
-                        }
-                        IconButton(onClick = onMoveNext) {
-                            Icon(
-                                Icons.AutoMirrored.Filled.ArrowForward,
-                                contentDescription = "次月へ移動",
-                                tint = MaterialTheme.colorScheme.onSurfaceVariant,
-                            )
-                        }
-                        IconButton(onClick = onEdit) {
-                            Icon(
-                                Icons.Default.Edit,
-                                contentDescription = "編集",
-                                tint = MaterialTheme.colorScheme.primary,
-                            )
-                        }
-                        IconButton(onClick = onDelete) {
-                            Icon(
-                                Icons.Default.Delete,
-                                contentDescription = "削除",
-                                tint = MaterialTheme.colorScheme.error,
-                            )
-                        }
-                    }
-                }
-            }
-
-            if (item.note.isNotBlank()) {
-                Spacer(modifier = Modifier.height(4.dp))
-                Text(
-                    text = item.note,
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                )
-            }
-
-            if (item.shares.isNotEmpty()) {
-                Spacer(modifier = Modifier.height(8.dp))
-                HorizontalDivider(color = MaterialTheme.colorScheme.outlineVariant)
-                Spacer(modifier = Modifier.height(8.dp))
-
-                for (share in item.shares) {
-                    val userName = users.find { it.uid == share.uid }?.displayName ?: share.uid.take(8)
-                    Row(
-                        modifier = Modifier.fillMaxWidth(),
-                        horizontalArrangement = Arrangement.SpaceBetween,
-                    ) {
-                        Text(text = userName, style = MaterialTheme.typography.bodyMedium)
-                        Text(
-                            text = formatYen(share.amount),
-                            style = MaterialTheme.typography.bodyMedium,
-                        )
-                    }
-                }
-
-                if (mismatch) {
-                    Spacer(modifier = Modifier.height(4.dp))
-                    Row(
-                        verticalAlignment = Alignment.CenterVertically,
-                        horizontalArrangement = Arrangement.spacedBy(4.dp),
-                    ) {
-                        Icon(
-                            Icons.Default.Warning,
-                            contentDescription = null,
-                            tint = MaterialTheme.colorScheme.error,
-                            modifier = Modifier.size(16.dp),
-                        )
-                        Text(
-                            text = "差額: ${formatYen(shareTotal - item.amount)}",
-                            style = MaterialTheme.typography.bodySmall,
-                            color = MaterialTheme.colorScheme.error,
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun ImportButtonText() {
-    Text(
-        text =
-            buildAnnotatedString {
-                append("前月の")
-                withStyle(SpanStyle(fontWeight = FontWeight.Bold)) {
-                    append("『毎月』")
-                }
-                append("項目をインポート")
-            },
     )
 }
